+++ tiger-rbel/src/main/java/de/gematik/rbellogger/writer/RbelXmlSerializer.java
@@ -6,6 +6,7 @@ package de.gematik.rbellogger.writer;
 
 import de.gematik.rbellogger.writer.RbelWriter.RbelWriterInstance;
 import de.gematik.rbellogger.writer.tree.RbelContentTreeNode;
+import de.gematik.rbellogger.writer.tree.RbelXmlElementToNodeConverter;
 import java.io.IOException;
 import java.io.StringWriter;
 import lombok.AllArgsConstructor;
@@ -35,20 +36,29 @@ public class RbelXmlSerializer implements RbelSerializer {
 
     private void addNode(RbelContentTreeNode treeNode, Branch parentBranch, RbelWriterInstance rbelWriter) {
         if (log.isTraceEnabled()) {
-            log.trace("converting xml node '{}'", treeNode.getContent() == null ? "<null>" : new String(treeNode.getContent()));
+            log.trace("converting xml node '{}'",
+                treeNode.getContent() == null ? "<null>"
+                    : new String(treeNode.getContent(), treeNode.getElementCharset()));
         }
-        final String key = treeNode.getKey().orElseThrow();
+        final String key = treeNode.getKey()
+            .orElseThrow(() -> new RbelSerializationException("Could not find key for " + treeNode));
 
         if (StringUtils.isEmpty(key)) {
             return;
-        } else if ("text".equals(key) || !treeNode.hasTypeOptional(RbelContentType.XML).orElse(true)) {
-            if (!(parentBranch instanceof Document)) {
-                parentBranch.add(new DefaultText(new String(rbelWriter.renderTree(treeNode).getContent(), treeNode.getElementCharset())));
+        } else {
+            final boolean isATextNode = "text".equals(key)
+                && !treeNode.hasTypeOptional(RbelContentType.XML).orElse(false)
+                && !treeNode.attributes().containsKey(RbelXmlElementToNodeConverter.IS_XML_ATTRIBUTE);
+            if (isATextNode) {
+                if (!(parentBranch instanceof Document)) {
+                    parentBranch.add(new DefaultText(
+                        new String(rbelWriter.renderTree(treeNode).getContent(), treeNode.getElementCharset())));
+                }
+                return;
+            } else if (treeNode.attributes().containsKey("isXmlAttribute") && parentBranch instanceof Element element) {
+                element.addAttribute(key, rbelWriter.renderTree(treeNode).getContentAsString());
+                return;
             }
-            return;
-        } else if (treeNode.attributes().containsKey("isXmlAttribute") && parentBranch instanceof Element element) {
-            element.addAttribute(key, rbelWriter.renderTree(treeNode).getContentAsString());
-            return;
         }
 
         final Element newElement = parentBranch.addElement(key);