+++ tiger-rbel/src/main/java/de/gematik/rbellogger/writer/tree/RbelXmlElementToNodeConverter.java
@@ -21,6 +21,8 @@ import org.apache.commons.lang3.tuple.Pair;
 @Slf4j
 public class RbelXmlElementToNodeConverter implements RbelElementToContentTreeNodeConverter {
 
+    public static final String IS_XML_ATTRIBUTE = "isXmlAttribute";
+
     @Override
     public boolean shouldConvert(RbelElement target) {
         return target.hasFacet(RbelXmlFacet.class);
@@ -43,7 +45,7 @@ public class RbelXmlElementToNodeConverter implements RbelElementToContentTreeNo
         for (RbelContentTreeNode node : converter.convertNode(entry.getValue(), entry.getKey(), context)) {
             // add attributes
             if (entry.getValue().hasFacet(RbelXmlAttributeFacet.class)) {
-                node.attributes().put("isXmlAttribute", "true");
+                node.attributes().put(IS_XML_ATTRIBUTE, "true");
             }
 
             // manage pulling up/down of text-nodes in mode-switches